requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.26.3
botocore==1.29.3
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
mailjet-rest==1.3.4
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
